+++ modelstudio-20260210/alibabacloud_modelstudio20260210/models/_create_workspace_request.py
@@ -7,8 +7,10 @@ from darabonba.model import DaraModel
 class CreateWorkspaceRequest(DaraModel):
     def __init__(
         self,
+        service_site: str = None,
         workspace_name: str = None,
     ):
+        self.service_site = service_site
         # This parameter is required.
         self.workspace_name = workspace_name
 
@@ -20,6 +22,9 @@ class CreateWorkspaceRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.service_site is not None:
+            result['serviceSite'] = self.service_site
+
         if self.workspace_name is not None:
             result['workspaceName'] = self.workspace_name
 
@@ -27,6 +32,9 @@ class CreateWorkspaceRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('serviceSite') is not None:
+            self.service_site = m.get('serviceSite')
+
         if m.get('workspaceName') is not None:
             self.workspace_name = m.get('workspaceName')