+++ modelstudio-20260210/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-04-27 Version: 1.0.1
+- Update API CreateWorkspace: add request parameters serviceSite.
+- Update API ListWorkspaces: add request parameters workspaceId.
+
+
 2026-03-31 Version: 1.0.0
 - Generated python 2026-02-10 for ModelStudio.